Change executable name in README instructions
The run command in the README.md file was corrected from `./healthchecker` to `./cmd`. This change is necessary to accurately reflect the correct name of the executable that is generated when the application is compiled with `go build`. This will eliminate any confusion for users trying to run the application.