Upgrade go-ole and update health-checker response
This commit upgrades go-ole from v1.2.6 to v1.3.0, providing an updated version of the package for better compatibility and performance. The golang.org/x/sys version was also added to go.sum providing additional required dependencies. Additionally, the health-checker functionality is updated to directly set the HTTP status to ServiceUnavailable if CPU or RAM load exceed the threshold, improving the accuracy of service status reporting. A test case is also added to simulate this scenario and ensure correct functioning.