Improved error handling and README documentation
A changeset has been applied to the way the application handles server errors in 'main.go'. This improvement means the application will now terminate if the server encounters a fatal error, enhancing the resilience of the service. Meanwhile, the 'Flag' field in README.md was updated to 'Flag/Environment Variable'. This change was made to better reflect the true functionality, as both flags and environment variables can be used.