Update dependencies and improve CPU load calculation
The dependencies in go.mod and go.sum have been updated, including the upgrade of github.com/shirou/gopsutil from version v3.21.11 to v3.23.9. Additionally, modification of CPU load calculations to provide a more accurate cumulative load across cores.

The change in CPU load calculation is needed as previously it only considered the load of a single core. The method cpu.Percent has a second parameter `percpu` that was set to false making it return just one overall percentage. However, in machines with multiple cores, this might lead to misleading results. Now it's changed to true making it return one value for each cpu and these values are summed to give the total load.