Refactor cpu load percentage calculation in health-checker
Simplified the calculation of cpu load percentage in health-checker module. Now instead of calculating the average of all CPU cores, the load of the first core is logged. Also, removed unnecessary variables and converted the cpu load to an integer by flooring it. This change makes the metrics easier to analyze, read, and use in health-check operations.