Updated the health-checker service path and extended shutdown timeout
Made changes in the health-checker. The service path for checking CPU and RAM load has been updated from '/' to '/check' to be more explicit. The shutdown timeout of the server has been extended from 3 seconds to 10 seconds for a smoother transition preventing premature termination of processes. CPU and RAM load calculations are now done with Context to ensure the process is cancellable and provides better control over go routines.