Refactor health-checker to improve project structure and efficiency
Removed main_test.go file and moved health-checker tests to new files in internal/handlers and internal/services. This separates the concerns of testing, making it easier to navigate and maintain. Renamed config/checker.go to internal/configs/checker.go and cmd/health-checker/main.go to cmd/main.go to reflect the internal structure of the application and make it more intuitive. Incorporated the services package to create a new Monitor type to handle CPU and RAM utilization, improving efficiency and readability of the program. Created new files for handlers and models to better organise the project. This optimisation helps to maintain the project easily and efficiently and adhere to microservice architecture principles.