Increase timeout context for Monitor Start unit test
This change increases the timeout context from 5 seconds to 10 seconds for the `Monitor_Start` unit test in `monitorService_test.go`. This will provide the test more leeway while ensuring the monitored service is working as expected. The adjustment resulted from instances where the test was failing due to timeout in a heavily loaded system.